+++ rotary_utils/utils.py
@@ -1,0 +1,70 @@
+#!/usr/bin/env python
+# utils.py
+# Utility function within rotary-utils
+# Copyright Jackson M. Tsuji and Lee H. Bergstrand 2024
+
+import logging
+import shlex
+import shutil
+import subprocess
+
+# Initialize the logger
+logger = logging.getLogger(__name__)
+
+
+def check_dependency(dependency_name: str):
+    """
+    Checks if a required shell dependency is present
+    :param dependency_name: name of the dependency
+    :return: path to the dependency
+    """
+
+    dependency_path = shutil.which(dependency_name)
+
+    if dependency_path is None:
+
+        logger.error(f'Dependency not found: {dependency_name}')
+        raise RuntimeError
+
+    return dependency_path
+
+
+def set_write_mode(append_log: bool):
+    """
+    Converts the boolean append_log to 'w' or 'a' write modes
+    :param append_log: boolean of whether to append to an existing log file (True) or to overwrite an existing log
+                       file (False)
+    :return: string of either 'a' (append mode) or 'w' (write mode)
+    """
+
+    if append_log is True:
+        write_mode = 'a'
+
+    elif append_log is False:
+        write_mode = 'w'
+
+    else:
+
+        logger.error(f'append_log should be a boolean True or False; you provided {append_log}')
+        raise ValueError
+
+    return write_mode
+
+
+def run_pipeline_subcommand(command_args, stdin=None, stdout=None, stderr=None, check=True):
+    """
+    Wrapper function for running subcommands.
+
+    :param command_args: The command line arguments of the subcommand (e.g., ['samtools', '-h'])
+    :param stdin: A subprocess.PIPE or None if stdin is not to be used.
+    :param stdout: Where to send stdout or None if stdout is not to be used.
+    :param stderr: Where to send stderr or None if stderr is not to be used.
+    :param check: Cause a runtime error if the subcommand fails.
+    :return: The output of the subcommand.
+    """
+    logger.debug(shlex.join(command_args))
+
+    if stdin:
+        stdin = stdin.stdout
+
+    return subprocess.run(command_args, check=check, input=stdin, stdout=stdout, stderr=stderr)
